requirements.txt
appdirs==1.4.3
Django==1.10.6
django-haystack==2.6.1
gunicorn==19.7.1
jieba==0.38
Markdown==2.6.8
packaging==16.8
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Whoosh==2.7.4
